Fix: Member엔티티 기본키 Long으로 수정

See merge request s09-fintech-finance-sub2/S09P22B109!20

+++ backend/economius/src/main/java/com/ssafy/economius/oauth/OAuthLoginService.java
@@ -2,11 +2,9 @@ package com.ssafy.economius.oauth;
 
 import com.ssafy.economius.game.entity.mysql.Member;
 import com.ssafy.economius.game.repository.mysql.MemberRepository;
-import lombok.Data;
+
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
-
-import java.time.LocalDateTime;
 
 @Service
 @RequiredArgsConstructor
@@ -17,12 +15,12 @@ public class OAuthLoginService {
 
     public AuthTokens login(OAuthLoginParams params) {
         OAuthInfoResponse oAuthInfoResponse = requestOAuthInfoService.request(params);
-        Integer memberId = findOrCreateMember(oAuthInfoResponse);
+        Long memberId = findOrCreateMember(oAuthInfoResponse);
         return authTokensGenerator.generate(memberId);
     }
 
-    private Integer findOrCreateMember(OAuthInfoResponse oAuthInfoResponse) {
-        Member member =  memberRepository.findByEmail(oAuthInfoResponse.getEmail());
+    private Long findOrCreateMember(OAuthInfoResponse oAuthInfoResponse) {
+        Member member = memberRepository.findByEmail(oAuthInfoResponse.getEmail());
         if (member != null) {
             return member.getMemberId();
         } else {
@@ -30,7 +28,7 @@ public class OAuthLoginService {
         }
     }
 
-    private Integer newMember(OAuthInfoResponse oAuthInfoResponse) {
+    private Long newMember(OAuthInfoResponse oAuthInfoResponse) {
         Member member = Member.builder()
                 .name(oAuthInfoResponse.getNickname())
                 .email(oAuthInfoResponse.getEmail())
